Updated the performance diagrams
Now shows performance diagrams of threshold 0 till 5.
Excluded diagrams for threshold of 10 and 30 (these did not show much because of the low scores)
Furthermore the FAR and POD are swapped when compared to previous plot
This is because the evaluation was done incorrectly. Target and predictions were swapped.
Now correct target and prediction is used to calculate FAR and POD